modify signup according to PR comments relates #23
